refactor: relocate tools/github/ + tools/shadow/ → src/Core.TypeScript/
tools/github/ → src/Core.TypeScript/github/ (12 TS + fixtures)
tools/shadow/ → src/Core.TypeScript/shadow/ (5 TS + AppleScript + launchd)

Import path fixes:
- Within github: ../../src/Core.TypeScript/X → ../X
- Within shadow: ../../src/Core.TypeScript/X → ../X
- Runtime paths (log file, plist templates, smoke-test REPO_ROOT)
- package.json bin/scripts paths

141 tests pass across 7 test files.

Continuing the tools/ → src/ migration. Only install/setup shell
scripts should remain in tools/ eventually.
