+++ package.json
@@ -18,11 +18,11 @@
     "hygiene:sort-tick-history": "bun ./tools/hygiene/sort-tick-history-canonical.ts",
     "hygiene:fix-markdown": "bun ./tools/hygiene/fix-markdown-md032-md026.ts",
     "hygiene:check-bash-retirement-inventory": "bun ./src/Core.TypeScript/hygiene/check-bash-retirement-inventory.ts --enforce",
-    "shadow": "bun tools/shadow/zeta-shadow.ts"
+    "shadow": "bun src/Core.TypeScript/shadow/zeta-shadow.ts"
   },
   "bin": {
     "ace": "src/Core.TypeScript/ace/ace.ts",
-    "zeta-shadow": "tools/shadow/zeta-shadow.ts"
+    "zeta-shadow": "src/Core.TypeScript/shadow/zeta-shadow.ts"
   },
   "devDependencies": {
     "@eslint/js": "10.0.1",
