+++ src/Core.TypeScript/github/fixtures/clean-armed-auto-merge.json
