+++ src/Core.TypeScript/github/fixtures/dirty-auto-merge-armed.json
