+++ src/Core.TypeScript/github/fixtures/github-status-all-operational.json
