+++ src/Core.TypeScript/github/fixtures/github-status-pr-degraded.json
