+++ src/Core.TypeScript/github/fixtures/non-required-failure-warning.json
