+++ src/Core.TypeScript/shadow/detect-grey-text.applescript
