+++ src/Core.TypeScript/shadow/get-frontmost-state.applescript
