+++ src/Core.TypeScript/shadow/outlet.ts
