+++ src/Core.TypeScript/shadow/restore-arrow.applescript
